+++ web-app/package.json
@@ -32,7 +32,7 @@
     "react-dom": "^16.3.2",
     "react-router-dom": "^4.2.2",
     "react-bootstrap": "0.32.1",
-    "react-bootstrap-navbar":"^1.1.0",
+    "react-bootstrap-navbar": "^1.1.0",
     "sweetalert2": "^7.19.2"
   }
 }
