+++ web-app/src/components/TakenLeave/style.css
