+++ web-app/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as RequestHoliday } from './RequestHoliday';
+export { default as BookedLeave } from './BookedLeave';
+export { default as TakenLeave } from './TakenLeave';
