+++ web-app/src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import App from './App'
 import Bootstrap from 'react-bootstrap';
-import Login from './Login/Login'
+import { Login } from './pages'
 
 ReactDOM.render(
   <Router>
