+++ web-app/src/pages/index.js
@@ -1,0 +1,2 @@
+export { default as Dashboard } from './Dashboard';
+export { default as Login } from './Login';
